Move generalworkflowIT tests
